Issue #11532 - Refactor common code into BaseForm
