+++ profile/modules/cp/modules/cp_taxonomy/src/Form/RemoveTermsFromNodeForm.php
@@ -2,105 +2,14 @@
 
 namespace Drupal\cp_taxonomy\Form;
 
-use Drupal\Core\Entity\EntityTypeManagerInterface;
-use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\Render\Renderer;
-use Drupal\Core\Session\AccountInterface;
-use Drupal\Core\TempStore\PrivateTempStoreFactory;
 use Drupal\Core\Url;
-use Drupal\cp_taxonomy\CpTaxonomyHelper;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 
 /**
  * Remove terms from node entities form.
  */
-class RemoveTermsFromNodeForm extends FormBase {
-
-  /**
-   * The array of entities to delete.
-   *
-   * @var array
-   */
-  protected $entityInfo = [];
-
-  /**
-   * The tempstore object.
-   *
-   * @var \Drupal\Core\TempStore\PrivateTempStore
-   */
-  protected $tempStore;
-
-  /**
-   * The entity type manager service.
-   *
-   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
-   */
-  protected $entityTypeManager;
-
-  /**
-   * The current user object.
-   *
-   * @var \Drupal\Core\Session\AccountInterface
-   */
-  protected $currentUser;
-
-  /**
-   * The entity type identifier.
-   *
-   * @var string
-   */
-  protected $entityTypeId;
-
-  /**
-   * Cp taxonomy helper.
-   *
-   * @var \Drupal\cp_taxonomy\CpTaxonomyHelper
-   */
-  protected $taxonomyHelper;
-
-  /**
-   * Renderer.
-   *
-   * @var \Drupal\Core\Render\Renderer
-   */
-  protected $renderer;
-
-  /**
-   * Constructs a DeleteMultiple form object.
-   *
-   * @param \Drupal\Core\TempStore\PrivateTempStoreFactory $temp_store_factory
-   *   The tempstore factory.
-   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $manager
-   *   The entity manager.
-   * @param \Drupal\Core\Session\AccountInterface $current_user
-   *   The current user object.
-   * @param \Drupal\cp_taxonomy\CpTaxonomyHelper $taxonomy_helper
-   *   Taxonomy Helper.
-   * @param \Drupal\Core\Render\Renderer $renderer
-   *   Taxonomy Helper.
-   */
-  public function __construct(PrivateTempStoreFactory $temp_store_factory, EntityTypeManagerInterface $manager, AccountInterface $current_user, CpTaxonomyHelper $taxonomy_helper, Renderer $renderer) {
-    $this->tempStore = $temp_store_factory->get('cp_taxonomy_remove_terms_node');
-    $this->entityTypeManager = $manager;
-    $this->currentUser = $current_user;
-    $this->taxonomyHelper = $taxonomy_helper;
-    $this->renderer = $renderer;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container) {
-    return new static(
-      $container->get('tempstore.private'),
-      $container->get('entity_type.manager'),
-      $container->get('current_user'),
-      $container->get('cp.taxonomy.helper'),
-      $container->get('renderer')
-    );
-  }
+class RemoveTermsFromNodeForm extends ManageTermsNodeFormBase {
 
   /**
    * {@inheritdoc}
@@ -113,6 +22,7 @@ class RemoveTermsFromNodeForm extends FormBase {
    * {@inheritdoc}
    */
   public function buildForm(array $form, FormStateInterface $form_state, $entity_type_id = NULL) {
+    $this->tempStore = $this->tempStoreFactory->get('cp_taxonomy_remove_terms_node');
     $form['#title'] = $this->t('Remove Terms from Content');
     $this->entityTypeId = $entity_type_id;
     $this->entityInfo = $this->tempStore->get($this->currentUser->id());
@@ -264,27 +174,6 @@ class RemoveTermsFromNodeForm extends FormBase {
     }
 
     $form_state->setRedirect('cp.content.collection');
-  }
-
-  /**
-   * Ajax callback for handling vocabulary depends terms selection.
-   */
-  public function getTermsAjaxCallback(array &$form, FormStateInterface $form_state) {
-    if ($selected_vocabulary = $form_state->getValue('vocabulary')) {
-      $vocabulary_storage = $this->entityTypeManager->getStorage('taxonomy_vocabulary');
-      $term_storage = $this->entityTypeManager->getStorage('taxonomy_term');
-      $vocab = $vocabulary_storage->load($selected_vocabulary);
-      // Get terms from selected vocabulary.
-      $terms = $term_storage->loadTree($vocab->id());
-      $options = [];
-      foreach ($terms as $term) {
-        $options[$term->tid] = $term->name;
-      }
-      $form['terms']['#options'] = $options;
-      // Return the prepared element.
-      return $form['terms'];
-    }
-    return [];
   }
 
 }
